fixes for broken tests
