+++ data-access-service/src/integrationTest/java/uk/gov/justice/laa/dstew/access/controller/application/GetApplicationTest.java
@@ -99,7 +99,12 @@ public class GetApplicationTest extends BaseHarnessTest {
         assertNoCacheHeaders(result);
         assertOK(result);
         ApplicationResponse expectedApplication = createApplication(savedApplication, proceeding, decision);
-        assertThat(actualApplication).isEqualTo(expectedApplication);
+        assertThat(actualApplication)
+            .usingRecursiveComparison()
+            .ignoringFields("lastUpdated", "version")
+            .isEqualTo(expectedApplication);
+        assertThat(actualApplication.getLastUpdated()).isNotNull();
+        assertThat(actualApplication.getVersion()).isEqualTo(1L);
     }
 
     @SmokeTest
