+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/repository/DecisionRepository.java
@@ -1,6 +1,5 @@
 package uk.gov.justice.laa.dstew.access.repository;
 
-import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
